+++ examples/remote-mdx/app/docs/[[...slug]]/page.tsx
@@ -1,8 +1,8 @@
 import type { Metadata } from 'next';
-import { DocsPage, DocsBody } from 'fumadocs-ui/page';
+import { DocsPage, DocsBody } from '@maximai/fumadocs-ui/page';
 import { notFound } from 'next/navigation';
 import { getDocs } from '@/app/source';
-import defaultComponents from 'fumadocs-ui/mdx';
+import defaultComponents from '@maximai/fumadocs-ui/mdx';
 import { resolveFile } from '@fumadocs/mdx-remote/github';
 import { compileMDX } from '@fumadocs/mdx-remote';